* Fixed some warnings in compile.

+++ src/arcemu-world/AchievementMgr.cpp
@@ -438,7 +438,7 @@ void AchievementMgr::UpdateAchievementCriteria(AchievementCriteriaTypes type, in
 							break;
 						case 247: // Make Love, Not Warcraft
 							{
-								Player* pTarget = objmgr.GetPlayer(selectedGUID);
+								Player* pTarget = objmgr.GetPlayer((uint32)selectedGUID);
 								if( pTarget && pTarget->IsPlayer() && pTarget->isDead() && isHostile(pTarget, GetPlayer()) )
 								{
 									UpdateCriteriaProgress(achievementCriteria, 1);
